ordered: parsnip Engines for Ordinal Regression Models

ordered provides engines for ordinal regression models for the parsnip package. The models may have cumulative, sequential, or adjacent-category structure, and in future these may be disaggregated into separate model types. A vignette will provide thorough illustrations of ordered functionality. See below for examples of fitting ordinal regression models with ordered.

Author

Maintainer: Jason Cory Brunson cornelioid@gmail.com (ORCID)

Authors:

Max Kuhn max@posit.co (ORCID)

Other contributors:

Posit Software PBC [copyright holder]

Examples

if (rlang::is_installed("MASS")) {

# Weighted sample

set.seed(561246)
house_sub <- MASS::housing |>
  dplyr::sample_n(size = 120, replace = TRUE, weight = Freq) |>
  subset(select = -Freq)
train_inds <- sample(120, 80)
house_train <- house_sub[train_inds, ]
house_test <- house_sub[-train_inds, ]

# Cumulative-link proportional-odds probit regression model

fit_cpop <- ordinal_reg() |>
  set_engine("polr") |>
  set_args(ordinal_link = "probit") |>
  fit(Sat ~ Infl + Type + Cont, data = house_train)
predict(fit_cpop, house_test, type = "prob")

if (rlang::is_installed("ordinalForest")) {

# Ordinal forest

fit_orf <- rand_forest(mode = "classification") |>
  set_engine("ordinalForest") |>
  set_args(nsets = 100, ntreefinal = 500, perffunction = "probability") |>
  fit(Sat ~ Infl + Type + Cont, data = house_train)
predict(fit_orf, house_test, type = "prob")

}
}
#> Warning: The argument `ntreefinal` cannot be manually modified and was removed.
#> # A tibble: 40 × 3
#>    .pred_Low .pred_Medium .pred_High
#>        <dbl>        <dbl>      <dbl>
#>  1     0.125        0.338      0.537
#>  2     0.512        0.194      0.294
#>  3     0.417        0.333      0.251
#>  4     0.512        0.194      0.294
#>  5     0.231        0.277      0.492
#>  6     0.436        0.329      0.236
#>  7     0.157        0.414      0.430
#>  8     0.235        0.214      0.550
#>  9     0.436        0.329      0.236
#> 10     0.125        0.338      0.537
#> # ℹ 30 more rows